细节优化.

+++ app/src/main/java/com/chocolate/nigerialoanapp/bean/response/FaceIdResponse.kt
@@ -1,5 +1,5 @@
 package com.chocolate.nigerialoanapp.bean.response
 
 class FaceIdResponse {
-    var face_id : Int? = null
+    var face_id : Long? = null
 }
